move tests to navigation
